Make webps first class citizens
Implements proper handling of the webp format, so that it is treated similar to
jpegs and pngs. Pre parsing of the header allows size and quality information
to be extracted without the need to decode the image. When encoding, the proper
parameters are passed to the encoder, as to respect quality levels and
losslessness