Return original image if late stage decoding fails
For compatibility reason, the original image is now returned if decoding fails
even after the initial checks. Because decoding effectively only happens right
before encoding, we can throw unexpected exceptions when `writeimage` is
called. To avoid this, we mimick the previous behavior, where failing to decode
could result in an empty list of operations, causing the original image to be
written