Check png transparency based on values in alpha channel
The `_ispngtransparent` function returns true if:

- Image is color_type 4 or 6, it has an alpha channel and this alpha
  channel has atleast one value that's not opaque (ie != 255). OpenCV is
used to extract this info.
- Image is color_type 0, 2, 3 and has tRNS chunk. To be precise presence
  of tRNS chunk alone is not sufficient to verify that image has a
transparent pixel, however currently the function returns true anyway
because in this case checking for actual transparency is more expensive
(OpenCV cannot be reliably used to expand tRNS chunk, so we'll need to
manually scan the image data to see if the color marked as transparent
by tRNS actually exists in image data).